It multiplies now
I tried libmatheval, which worked, but what I had before was actually better;
it did parentheses, and functions, and other stuff I know I'll want down the
road (not to mention its variables only allowed X-Z).

What I have here works, so now I just need to finish setting it up.  It only does
one (so 3 * 3 * 3 gives me 9 * 3, not 81) but that is easy to fix.

+++ new/replace_with_float.c
@@ -1,6 +1,6 @@
 #include "main.h"
 
-void replace_with_float(char* line, uint16_t from, uint16_t to, float value) {
+void replace_with_float(char* line, size_t from, size_t to, float value) {
         size_t i;
 	char temp[LINE_SIZE];
         memset(temp, 0, LINE_SIZE);
